refactor: split single huge main.ts to multiple files
comment 相关的逻辑比较耦合，于是就没有继续拆
