+++ frontend/lib/const.ts
@@ -1,0 +1,4 @@
+export let apiEndpoint = "/api";
+export const setApiEndpoint = (endpoint: string) => {
+  apiEndpoint = endpoint;
+};
